Support all datatypes for Xtensa softmax implementation.
 * This allows the micro_speech and person_detect examples to be usable
   with the optimized xtensa kernels.

 * enabled all the softmax kernel test cases for Xtensa (since we have a
   fallback to the reference kernels).

 * The keyword_benchmark specific variant is moved to its own .cc file
   since that allows the Xtensa linker to properly drop unsused symbols
   and ensures that the binary size for the keyword_benchmark is
   unchanged.

Manually tested that the following tests pass:
```
make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa TARGET_ARCH=fusion_f1 OPTIMIZED_KERNEL_DIR=xtensa XTENSA_CORE=F1_190305_swupgrade test_person_detection_test_int8 -j8
make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa TARGET_ARCH=fusion_f1 OPTIMIZED_KERNEL_DIR=xtensa XTENSA_CORE=F1_190305_swupgrade test_micro_speech_test -j8
make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa TARGET_ARCH=fusion_f1 OPTIMIZED_KERNEL_DIR=xtensa XTENSA_CORE=F1_190305_swupgrade test_kernel_softmax_test -j8
```

Confirmed that the binary size for the keyword_benchmark is unchanged (relative to tip of tree):
```
make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa TARGET_ARCH=fusion_f1 OPTIMIZED_KERNEL_DIR=xtensa XTENSA_CORE=F1_190305_swupgrade keyword_benchmark -j8 BUILD_TYPE=release
xt-size tensorflow/lite/micro/tools/make/gen/xtensa_fusion_f1_release/bin/keyword_benchmark
```

Gives:
```
   text	   data	    bss	    dec	    hex	filename
  70368	  41140	  24856	 136364	  214ac	tensorflow/lite/micro/tools/make/gen/xtensa_fusion_f1_release/bin/keyword_benchmark
```
